matrixop from<> and to<> are not standards-compliant. Too bad. Will redo in the future. Fixed all references to avoid.
genparams-qt updated to remove baseParFile ref in ddparGenerator. It was removed.

Need to get rid of C4996 warning with boost uuid.


git-svn-id: file:///mnt/svn/rtmath@677 1d52797c-35fc-48e8-8e80-2db72aa187e2

+++ rtmath/rtmath-apps/ddscat-genparams/main.cpp
@@ -123,7 +123,7 @@ int main(int argc, char** argv)
 			debug::debug_preamble(vout);
 			vout << endl << endl;
 			vout << "Command line:\n";
-			for (size_t i=0; i<argc;i++)
+			for (size_t i=0; i<(size_t)argc;i++)
 				vout << argv[i] << " ";
 			vout << endl;
